requirements.txt
Django==3.2.10
django-filter==21.1
django-extensions==3.1.5
djangorestframework==3.13.1
django-cors-headers
drf-yasg==1.20.0
psycopg2-binary==2.9.2
requests==2.26.0
gevent==21.1.2
gunicorn

